use matrices for week 5 q3
git-svn-id: http://svn.codalism.com/svn/comp90042-2014s1/xxu@1581 1e191c3d-2cf9-4e68-b3ea-b23acec113d7

+++ workshops/lib/store/classif.py
@@ -17,7 +17,7 @@ class ClassifDataStore:
             doc_labels = sorted(classif_data.doc_to_labels.get(doc_id, []))
             row = np.zeros(len(doc_labels))
             col = np.array(list(self.label_indexes[label] for label in doc_labels))
-            data = np.ones(len(doc_labels))
+            data = np.ones(len(doc_labels), dtype=np.int)
             labels_rows.append(csr_matrix((data, (row, col)), shape=(1, len(self.labels))))
         self.labels_matrix = vstack(labels_rows).tocsr()
